Generalize the logics behind the PLT evaluation on BGG
